# frontend/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Semenq API specification - AI-powered medicine accessibility platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pharmacies:
    get:
      operationId: listPharmacies
      tags: [pharmacies]
      summary: List pharmacies
      parameters:
        - name: city
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
        - name: lat
          in: query
          schema: { type: number }
        - name: lng
          in: query
          schema: { type: number }
        - name: radius
          in: query
          schema: { type: number }
      responses:
        "200":
          description: List of pharmacies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pharmacy"
  /pharmacies/{id}:
    get:
      operationId: getPharmacy
      tags: [pharmacies]
      summary: Get pharmacy details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Pharmacy details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pharmacy"
  /pharmacies/{id}/inventory:
    get:
      operationId: getPharmacyInventory
      tags: [pharmacies]
      summary: Get pharmacy inventory
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Inventory list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
...
components:
  schemas:
...
    Pharmacy:
      type: object
      required: [id, name, city, state, address, phone, isVerified, rating]
      properties:
        id: { type: integer }
        name: { type: string }
        ownerName: { type: string }
        address: { type: string }
        city: { type: string }
        state: { type: string }
        pincode: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        lat: { type: number }
        lng: { type: number }
        isVerified: { type: boolean }
        rating: { type: number }
        reviewCount: { type: integer }
        openTime: { type: string }
        closeTime: { type: string }
        offersCourier: { type: boolean }
        licenseNumber: { type: string }
        totalInventory: { type: integer }
    InventoryItem:
      type: object
      required: [id, medicineId, medicineName, quantity, price, expiryDate, stockStatus]
      properties:
        id: { type: integer }
        medicineId: { type: integer }
        medicineName: { type: string }
        genericName: { type: string }
        quantity: { type: integer }
        price: { type: number }
        mrp: { type: number }
        expiryDate: { type: string }
        batchNumber: { type: string }
        stockStatus: { type: string, enum: [inStock, lowStock, outOfStock] }
        reorderLevel: { type: integer }
        lastRestocked: { type: string }
...